[Edit] Актуализация тестов

+++ Tests/DatalakeServer.Tests/Tests/CollectorsTests/CollectFromLocalInopc.cs
@@ -1,14 +1,16 @@
 ﻿using DatalakeApiClasses.Enums;
 using DatalakeApiClasses.Models.Sources;
 using DatalakeApiClasses.Models.Tags;
+using DatalakeApiClasses.Models.Users;
 using DatalakeApiClasses.Models.Values;
+using DatalakeServer.Constants;
 using DatalakeServer.Services.Receiver.Models.Inopc;
-using DatalakeServer.Tests.Attributes;
-using DatalakeServer.Tests.Extensions;
+using DatalakeServer.TestRunner.Attributes;
+using DatalakeServer.TestRunner.Extensions;
 using Microsoft.Extensions.Logging;
 using System.Net.Http.Json;
 
-namespace DatalakeServer.Tests.Tests.CollectorsTests;
+namespace DatalakeServer.TestRunner.Tests.CollectorsTests;
 
 public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>>
 {
@@ -30,10 +32,26 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 	const string TestItemPath = "Test.TestDevice.Num1";
 	const string TestTagName = "TestTag1";
 
+	public async Task<(string, string)> Authorization()
+	{
+		var request = new UserLoginPass { Name = "admin", Password = "admin" };
+		var response = await _httpClient.PostAsJsonAsync("api/users/auth", request);
+
+		var token = response.Headers.GetValues(AuthConstants.TokenHeader).First();
+		var name = response.Headers.GetValues(AuthConstants.NameHeader).First();
+
+		return (token, name);
+	}
+
 	[Fact, Priority(0)]
-	public void T000_Prepare()
+	public async Task T000_PrepareAsync()
 	{
 		Assert.True(true);
+
+		var (token, name) = await Authorization();
+
+		Assert.NotNull(token);
+		Assert.NotNull(name);
 	}
 
 	[Fact, Priority(1)]
@@ -53,6 +71,10 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 	[Fact, Priority(2)]
 	public async Task T002_CreateSourceForLocalInopc()
 	{
+		var (token, name) = await Authorization();
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.TokenHeader, token);
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.NameHeader, name);
+
 		var body = new SourceInfo
 		{
 			Name = TestSourceName,
@@ -67,13 +89,17 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 	[Fact, Priority(3)]
 	public async Task T003_GetItemsListForTestSource()
 	{
+		var (token, name) = await Authorization();
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.TokenHeader, token);
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.NameHeader, name);
+
 		var sources = await _httpClient.GetAsync<SourceInfo[]>("api/sources");
 		Assert.NotNull(sources);
 
 		var testSource = sources.FirstOrDefault(x => x.Name == TestSourceName);
 		Assert.NotNull(testSource);
 
-		var items = await _httpClient.GetAsync<SourceEntryInfo[]>($"api/sources/{testSource.Id}/tags");
+		var items = await _httpClient.GetAsync<SourceEntryInfo[]>($"api/sources/{testSource.Id}/items-and-tags");
 		Assert.True(items.Length >= 3);
 		Assert.Contains(items, x => x.ItemInfo?.Path == TestItemPath);
 	}
@@ -81,6 +107,10 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 	[Fact, Priority(4)]
 	public async Task T004_CreateTestTagForTestSource()
 	{
+		var (token, name) = await Authorization();
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.TokenHeader, token);
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.NameHeader, name);
+
 		var sources = await _httpClient.GetAsync<SourceInfo[]>("api/sources");
 		Assert.NotNull(sources);
 
@@ -96,12 +126,27 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 		});
 
 		Assert.True(id > 0);
+
+		await _httpClient.PutAsync("api/tags/" + id, new TagUpdateRequest
+		{
+			Name = TestTagName,
+			IntervalInSeconds = 0,
+			Type = TagType.Number,
+			SourceId = testSource.Id,
+			SourceItem = TestItemPath,
+			IsScaling = false,
+			SourceType = SourceType.Inopc,
+		});
 	}
 
 	[Fact, Priority(5)]
 	public async Task T005_EnsureValueOfTestTagsUpdatesTwice()
 	{
-		await Task.Delay(TimeSpan.FromSeconds(1));
+		await Task.Delay(TimeSpan.FromSeconds(3));
+
+		var (token, name) = await Authorization();
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.TokenHeader, token);
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.NameHeader, name);
 
 		object? storedValue = null;
 
@@ -135,20 +180,27 @@ public class CollectFromLocalInopc : IClassFixture<TestingWebAppFactory<Program>
 	[Fact, Priority(int.MaxValue)]
 	public async Task T999_RemoveTestObjects()
 	{
+		var (token, name) = await Authorization();
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.TokenHeader, token);
+		_httpClient.DefaultRequestHeaders.Add(AuthConstants.NameHeader, name);
+
 		var sources = await _httpClient.GetAsync<SourceInfo[]>("api/sources");
 		Assert.NotNull(sources);
 
 		var testSource = sources.FirstOrDefault(x => x.Name == TestSourceName);
-		Assert.NotNull(testSource);
-
-		await _httpClient.DeleteAsync("api/sources/" + testSource.Id);
+		if (testSource != null)
+		{
+			await _httpClient.DeleteAsync("api/sources/" + testSource.Id);
+		}
 
 		var tags = await _httpClient.GetAsync<TagInfo[]>("api/tags");
 		Assert.NotNull(tags);
 
 		var testTag = tags.FirstOrDefault(x => x.Name == TestTagName);
-		Assert.NotNull(testTag);
-
-		await _httpClient.DeleteAsync("api/tags/" + testTag.Id);
+		if (testTag != null)
+		{
+			Assert.NotNull(testTag);
+			await _httpClient.DeleteAsync("api/tags/" + testTag.Id);
+		}
 	}
 }
